requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via
    #   cohere
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annoy==1.17.3
    # via -r requirements.in
anyio==3.7.1
    # via langchain
appnope==0.1.3
    # via ipython
astor==0.8.1
    # via pandasai
asttokens==2.4.0
    # via stack-data
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
backcall==0.2.0
    # via ipython
backoff==2.2.1
    # via cohere
beautifulsoup4==4.12.2
    # via bs4
bs4==0.0.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via pip-tools
cohere==4.27
    # via -r requirements.in
contourpy==1.1.1
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.6.0
    # via langchain
decorator==5.1.1
    # via ipython
duckdb==0.8.1
    # via pandasai
exceptiongroup==1.1.3
    # via
    #   anyio
    #   ipython
executing==1.2.0
    # via stack-data
fastavro==1.8.2
    # via cohere
fonttools==4.42.1
    # via matplotlib
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via
    #   build
    #   cohere
importlib-resources==6.1.0
    # via matplotlib
ipython==8.15.0
    # via pandasai
jedi==0.19.0
    # via ipython
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
kiwisolver==1.4.5
    # via matplotlib
langchain==0.0.300
    # via -r requirements.in
langsmith==0.0.40
    # via langchain
marshmallow==3.20.1
    # via dataclasses-json
matplotlib==3.8.0
    # via pandasai
matplotlib-inline==0.1.6
    # via ipython
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.6
    # via langchain
numpy==1.26.0
    # via
    #   contourpy
    #   langchain
    #   matplotlib
    #   numexpr
    #   pandas
    #   scipy
openai==0.27.10
    # via pandasai
packaging==23.1
    # via
    #   build
    #   marshmallow
    #   matplotlib
pandas==1.5.3
    # via pandasai
pandasai==1.2.6
    # via -r requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.1
    # via matplotlib
pip-tools==7.3.0
    # via -r requirements.in
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pydantic==1.10.12
    # via
    #   langchain
    #   langsmith
    #   pandasai
pygments==2.16.1
    # via ipython
pyparsing==3.1.1
    # via matplotlib
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pandasai
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via langchain
requests==2.31.0
    # via
    #   cohere
    #   langchain
    #   langsmith
    #   openai
scipy==1.11.2
    # via pandasai
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.0
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.21
    # via
    #   langchain
    #   pandasai
stack-data==0.6.2
    # via ipython
tenacity==8.2.3
    # via langchain
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
tqdm==4.66.1
    # via openai
traitlets==5.10.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.8.0
    # via
    #   ipython
    #   pydantic
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.5
    # via
    #   cohere
    #   requests
wcwidth==0.2.6
    # via prompt-toolkit
wheel==0.41.2
    # via pip-tools
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
